+++ rbx_types/CHANGELOG.md
@@ -1,7 +1,17 @@
 # rbx_types Changelog
 
 ## Unreleased Changes
+
+## 1.4.2 (2022-06-12)
+* `Variant::String` now encodes correctly inside of `Attributes`.
+
+## 1.4.1 (2022-06-12)
+* `Attributes` can now decode from an empty buffer.
+* `Attributes` now encodes as an empty buffer when empty.
+
+## 1.4.0 (2022-06-05)
 * Added `Attributes::with` for creating attributes in code more easily.
+* Implemented `Hash` for `VariantTy`.
 
 ## 1.3.0 (2021-10-11)
 * Implemented `Tags`. ([#199])